apps/ai/requirements.txt
# Framework
fastapi>=0.115
uvicorn[standard]==0.30.0
httpx==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pgvector>=0.3

# Auth
python-jose[cryptography]>=3.3
bcrypt>=4.0

# Redis
redis>=5.0

# AI
openai>=1.0
anthropic>=0.34

# GraphQL
strawberry-graphql[asgi]>=0.240

# File parsing
openpyxl>=3.1
pdfplumber>=0.11
python-docx>=1.1
